+++ tests/utils/urls.py
@@ -1,6 +1,6 @@
 from django.conf import settings
 from django.contrib import admin
-from django.urls import include, re_path
+from django.urls import include, path
 from django.views.static import serve
 
 
@@ -8,9 +8,11 @@ admin.autodiscover()
 admin_urls = admin.site.urls
 
 urlpatterns = [
-    re_path(r'^media/(?P<path>.*)$', serve,
-            {'document_root': settings.MEDIA_ROOT, 'show_indexes': True}),
-    re_path(r'^admin/', admin_urls),
-    re_path(r'^', include('filer.server.urls')),
-    re_path(r'^filer/', include('filer.urls')),
+    path('media/my-preferred-base-url-for-source-files/<path:path>', serve,
+         {'document_root': settings.MEDIA_ROOT, 'show_indexes': True}),
+    path('media/my-preferred-base-url-for-thumbnails/<path:path>', serve,
+         {'document_root': settings.MEDIA_ROOT, 'show_indexes': True}),
+    path('admin/', admin_urls),
+    path('', include('filer.server.urls')),
+    path('filer/', include('filer.urls')),
 ]